+++ tests/test_action_guest_dir_delete.py
@@ -33,5 +33,7 @@ class DeleteDirectoryInGuestTestCase(VsphereBaseActionTestCase):
         action.si_content.guestOperationsManager.fileManager = mockFileMgr
         result = action.run(vm_id='vm-12345', username='u', password='p',
                             guest_directory='/tmp/foo', recursive=True)
-        mockFileMgr.DeleteDirectoryInGuest.assert_called_once()
+        mockFileMgr.DeleteDirectoryInGuest.assert_called_once_with(
+            mock_vm, action.guest_credentials, '/tmp/foo', True
+        )
         self.assertEqual(result, None)