+++ tests/test_action_guest_file_create.py
@@ -34,5 +34,7 @@ class CreateTemporaryFileInGuestTestCase(VsphereBaseActionTestCase):
         action.si_content.guestOperationsManager.fileManager = mockFileMgr
         result = action.run(vm_id='vm-12345', username='u', password='p',
                             guest_directory='d', prefix='p', suffix='s')
-        mockFileMgr.CreateTemporaryFileInGuest.assert_called_once()
+        mockFileMgr.CreateTemporaryFileInGuest.assert_called_once_with(
+            mock_vm, action.guest_credentials, 'p', 's', 'd'
+        )
         self.assertEqual(result, '/tmp/foo.txt')